refactor(pms): route procurement write reads through integration client
