+++ app/procurement/repos/receive_po_line_repo.py
@@ -5,22 +5,19 @@ from typing import Optional, Tuple
 from fastapi import HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.pms.export.items.services.item_read_service import ItemReadService
-from app.pms.export.uoms.contracts.uom import PmsExportUom
-from app.pms.export.uoms.services.uom_read_service import PmsExportUomReadService
+from app.integrations.pms.contracts import PmsExportUom
+from app.integrations.pms.inprocess_client import InProcessPmsReadClient
 
 
 async def load_item_expiry_policy(session: AsyncSession, *, item_id: int) -> str:
-    svc = ItemReadService(session)
-    policy = await svc.aget_policy_by_id(item_id=int(item_id))
+    policy = await InProcessPmsReadClient(session).get_item_policy(item_id=int(item_id))
     if policy is None:
         return "NONE"
     return str(policy.expiry_policy or "NONE").upper()
 
 
 async def load_item_lot_source_policy(session: AsyncSession, *, item_id: int) -> str:
-    svc = ItemReadService(session)
-    policy = await svc.aget_policy_by_id(item_id=int(item_id))
+    policy = await InProcessPmsReadClient(session).get_item_policy(item_id=int(item_id))
     if policy is None:
         return "INTERNAL_ONLY"
     return str(policy.lot_source_policy or "INTERNAL_ONLY").upper()
@@ -40,7 +37,7 @@ async def require_item_uom_ratio_to_base(
     if uom_id <= 0:
         raise HTTPException(status_code=400, detail="uom_id 必须为正整数")
 
-    row = await PmsExportUomReadService(session).aget_by_id(item_uom_id=int(uom_id))
+    row = await InProcessPmsReadClient(session).get_uom(item_uom_id=int(uom_id))
     if row is None or int(row.item_id) != int(item_id):
         raise HTTPException(
             status_code=400,
